перелопатил код создания дорог

+++ Traffic project/src/main/java/ru/nsu/fit/traffic/jsonParse/NodeJsonStruct.java
@@ -5,14 +5,14 @@ import ru.nsu.fit.traffic.model.TrafficLight;
 import java.util.List;
 
 public class NodeJsonStruct {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private List<Integer> roadsIn;
     private List<Integer> roadsOut;
     private boolean isSpawner;
     private TrafficLight trafficLight;
 
-    public NodeJsonStruct(int x, int y, List<Integer> roadsFrom, List<Integer> roadsTo, boolean isSpawner, TrafficLight trafficLight) {
+    public NodeJsonStruct(double x, double y, List<Integer> roadsFrom, List<Integer> roadsTo, boolean isSpawner, TrafficLight trafficLight) {
         this.x = x;
         this.y = y;
         this.roadsIn = roadsFrom;
